AUT-5652: Extract LambdaPauseHelper to shared module
The "pause between invocations" logic in `BulkEmailBatchPauseHelper` is needed by the upcoming inactive account data export self-invocation feature. It has therefore been extracted to a generic shared helper.

The interrupt unit test performs a no-op call before testing on an interrupted thread because the Log4j LogManager appears to fail permanently if the thread is interrupted during first class load.
